requirements.txt
wheel
numpy
pygame
simplejson
requests
pillow

